LayoutTests/fast/forms/file/entries-api/webkitEntries-nonascii-folder-name-expected.txt

OPEN FILE PANEL
Tests webkitEntries when the path includes a folder with a non-ascii character in its name.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS entries.length is 1
PASS entry.name is "test.txt"
PASS entry.fullPath is "/test.txt"
PASS entry.isFile is true
PASS contentResult is "TEST\n"
PASS successfullyParsed is true

TEST COMPLETE
